Cacheクラス修正

+++ core/WebStream/Module/Cache.php
@@ -1,18 +1,17 @@
 <?php
 namespace WebStream\Module;
 
+use WebStream\Exception\IOException;
+
 /**
  * Cacheクラス
  * @author Ryuichi Tanaka
  * @since 2011/10/08
- * @version 0.4
+ * @version 0.4.1
  */
 class Cache
 {
     use Utility;
-
-    const CACHE_DIR_WIN = "C:\\Windows\\Temp";
-    const CACHE_DIR_UNIX = "/tmp/";
 
     /** キャッシュの保存ディレクトリ */
     private $savePath;
@@ -25,11 +24,7 @@ class Cache
     {
         $this->savePath = $savePath;
         if ($this->savePath === null) {
-            if (PHP_OS === "WIN32" || PHP_OS === "WINNT") {
-                $this->savePath = self::CACHE_DIR_WIN;
-            } else {
-                $this->savePath = self::CACHE_DIR_UNIX;
-            }
+            $this->savePath = $this->getTemporaryDirectory();
         }
     }
 
@@ -42,7 +37,7 @@ class Cache
     {
         $cache = $this->cache($id);
         if ($cache !== null) {
-            $path = $this->savePath . $id . '.cache';
+            $path = $this->savePath . "/" . $id . '.cache';
             $cachePath = realpath($path);
             Logger::info("Get cache: ${cachePath}");
 
@@ -76,16 +71,20 @@ class Cache
      */
     private function cache($id)
     {
-        $path = $this->savePath . $id . '.cache';
+        $path = $this->savePath . "/" . $id . '.cache';
         $cachePath = realpath($path);
         if ($cachePath !== false && is_file($cachePath)) {
             $data = $this->decode(file_get_contents($cachePath));
-            // 期限切れのキャッシュは削除
-            if (time() > $data["time"] + $data["ttl"]) {
-                Logger::warn("Expired cache: ${cachePath}");
-                unlink($cachePath);
 
-                return null;
+            // 制限を確認
+            if ($data["ttl"] !== -1) {
+                // 期限切れのキャッシュは削除
+                if (time() > $data["time"] + $data["ttl"]) {
+                    Logger::warn("Expired cache: ${cachePath}");
+                    unlink($cachePath);
+
+                    return null;
+                }
             }
 
             return $data;
@@ -98,10 +97,11 @@ class Cache
      * キャッシュを保存する
      * @param string キャッシュID
      * @param object 保存データ
-     * @param integer キャッシュ保存時間
+     * @param integer キャッシュ保存時間(デフォルト無期限)
      * @param boolean 上書きするかどうか
+     * @return boolean 保存結果
      */
-    public function save($id, $data, $ttl = 60, $overwrite = false)
+    public function save($id, $data, $ttl = -1, $overwrite = false)
     {
         $content = [
             "time" => time(),
@@ -115,19 +115,15 @@ class Cache
             // キャッシュファイルがない場合またはキャッシュファイルが存在するが、
             // 上書きする場合はキャッシュを新規作成する
             if (!is_file($cachePath) || (is_file($cachePath) && $overwrite === true)) {
-                try {
-                    $result = file_put_contents($cachePath, $this->encode($content));
-                    // ファイルが書き込めた場合
-                    if ($result !== false) {
-                        Logger::info("Create cache: ${cachePath}");
-                        // キャッシュファイルのパーミッションを777にする
-                        @chmod($cachePath, 0777);
+                $result = file_put_contents($cachePath, $this->encode($content), LOCK_EX);
+                if ($result !== false) { // ファイルが書き込めた場合
+                    Logger::info("Create cache: ${cachePath}");
+                    // キャッシュファイルのパーミッションを777にする
+                    @chmod($cachePath, 0777);
 
-                        return true;
-                    }
-                    Logger::error("Can't create cache: ${cachePath}");
-                } catch (\Exception $e) {
-                    Logger::error($e->getMessage(), $e->getTraceAsstring());
+                    return true;
+                } else { // ファイル書き込みに失敗した場合
+                    throw new IOException("Can't create cache: " . $cachePath);
                 }
             }
         } else {
